practica3-main/requirements.txt
paho-mqtt
discord.py
pytest